+++ aa-sdk/core/src/actions/smartAccount/internal/initUserOperation.ts
@@ -62,7 +62,8 @@ export async function _initUserOperation<
 
   const signature = account.getDummySignature();
 
-  const nonce = account.getAccountNonce(overrides?.nonceKey);
+  const nonce =
+    overrides?.nonce ?? account.getAccountNonce(overrides?.nonceKey);
 
   const struct =
     entryPoint.version === "0.6.0"