+++ account-kit/java/signer-sdk/src/main/java/com/alchemy/aa/core/exceptions/NoInjectedBundleException.java
@@ -1,0 +1,6 @@
+package com.alchemy.aa.core.exceptions;
+
+public class NoInjectedBundleException extends RuntimeException {
+
+  public NoInjectedBundleException() {}
+}