+++ examples/react-native-expo-example/metro.config.js
@@ -8,6 +8,14 @@ const monorepoRoot = path.resolve(projectRoot, "../..");
 
 const config = getDefaultConfig(projectRoot);
 
+// Add aliases for file-system import based modules
+const ALIASES = {
+	"@noble/hashes/crypto": path.resolve(
+		monorepoRoot,
+		"node_modules/@noble/hashes/crypto.js"
+	),
+};
+
 // 1. Watch all files within the monorepo
 config.watchFolders = [projectRoot, monorepoRoot];
 
@@ -21,14 +29,30 @@ config.resolver.nodeModulesPaths = [
 // Force Metro to resolve (sub)dependencies only from the `nodeModulesPaths`
 config.resolver.disableHierarchicalLookup = true;
 
+// Default to file-based module resolution for file-system import based modules
+config.resolver.resolveRequest = (context, moduleName, platform) => {
+	if (ALIASES[moduleName]) {
+		return {
+			filePath: ALIASES[moduleName],
+			type: "sourceFile",
+		};
+	}
+	return context.resolveRequest(context, moduleName, platform);
+};
+
 config.resolver.extraNodeModules = {
 	...config.resolver.extraNodeModules,
-	...require("node-libs-react-native"),
-	crypto: require.resolve("crypto-browserify"),
+	crypto: require.resolve('crypto-browserify'),
 	stream: require.resolve("stream-browserify"),
 };
 
 // Important to allow importing package exports
 config.resolver.unstable_enablePackageExports = true;
+
+config.resolver.unstable_conditionNames = [
+	"browser",
+	"require",
+	"react-native",
+];
 
 module.exports = config;