+++ examples/ui-demo/src/components/icons/twitch.tsx
@@ -1,0 +1,31 @@
+export const TwitchIcon = ({ ...props }: React.SVGProps<SVGSVGElement>) => (
+  <svg xmlns="http://www.w3.org/2000/svg" width={20} height={20} {...props}>
+    <path
+      d="m16.547 9.344-2.617 2.62h-2.621l-2.29 2.29v-2.29H6.07v-9.82h10.477Zm0 0"
+      style={{
+        stroke: "none",
+        fillRule: "nonzero",
+        fill: "#fff",
+        fillOpacity: 1,
+      }}
+    />
+    <path
+      d="M5.418.832 2.145 4.105v11.79H6.07v3.273l3.274-3.273h2.62L17.856 10V.832Zm11.129 8.512-2.617 2.62h-2.621l-2.29 2.294v-2.293H6.07v-9.82h10.477Zm0 0"
+      style={{
+        stroke: "none",
+        fillRule: "nonzero",
+        fill: "#9146ff",
+        fillOpacity: 1,
+      }}
+    />
+    <path
+      d="M14.582 4.434h-1.309v3.93h1.309ZM10.98 4.434H9.672v3.93h1.308Zm0 0"
+      style={{
+        stroke: "none",
+        fillRule: "nonzero",
+        fill: "#9146ff",
+        fillOpacity: 1,
+      }}
+    />
+  </svg>
+);